+++ pkg/responsebuilder/algorithms_response_builder_test.go
@@ -63,9 +63,9 @@ func TestAlgorithmsResponse(t *testing.T) {
 				}
 
 				if tc.ComponentErrorMessages[i] != "" {
-					assert.NotNil(t, purl.ErrorMessage)
-					if purl.ErrorMessage != nil {
-						assert.Equal(t, tc.ComponentErrorMessages[i], *purl.ErrorMessage)
+					assert.NotNil(t, purl.InfoMessage)
+					if purl.InfoMessage != nil {
+						assert.Equal(t, tc.ComponentErrorMessages[i], *purl.InfoMessage)
 					}
 				}
 			}
@@ -105,9 +105,9 @@ func TestAlgorithmsResponse(t *testing.T) {
 				}
 
 				if tc.ComponentErrorMessages[i] != "" {
-					assert.NotNil(t, component.ErrorMessage)
-					if component.ErrorMessage != nil {
-						assert.Equal(t, tc.ComponentErrorMessages[i], *component.ErrorMessage)
+					assert.NotNil(t, component.InfoMessage)
+					if component.InfoMessage != nil {
+						assert.Equal(t, tc.ComponentErrorMessages[i], *component.InfoMessage)
 					}
 				}
 			}