+++ pkg/responsebuilder/hints_in_range_response_builder_test.go
@@ -135,9 +135,9 @@ func TestHintsInRangeResponse(t *testing.T) {
 				assert.Equal(t, tc.ComponentHintsCount[i], len(purl.Hints))
 
 				if tc.ComponentErrorMessages[i] != "" {
-					assert.NotNil(t, purl.ErrorMessage)
-					if purl.ErrorMessage != nil {
-						assert.Equal(t, tc.ComponentErrorMessages[i], *purl.ErrorMessage)
+					assert.NotNil(t, purl.InfoMessage)
+					if purl.InfoMessage != nil {
+						assert.Equal(t, tc.ComponentErrorMessages[i], *purl.InfoMessage)
 					}
 				}
 			}
@@ -174,9 +174,9 @@ func TestHintsInRangeResponse(t *testing.T) {
 				assert.Equal(t, tc.ComponentHintsCount[i], len(component.Hints))
 
 				if tc.ComponentErrorMessages[i] != "" {
-					assert.NotNil(t, component.ErrorMessage)
-					if component.ErrorMessage != nil {
-						assert.Equal(t, tc.ComponentErrorMessages[i], *component.ErrorMessage)
+					assert.NotNil(t, component.InfoMessage)
+					if component.InfoMessage != nil {
+						assert.Equal(t, tc.ComponentErrorMessages[i], *component.InfoMessage)
 					}
 				}
 			}